requirements.txt
torch>=2.0.0
torchvision
transformers>=4.45.0
accelerate>=0.26.0
bitsandbytes>=0.43.0
qwen-vl-utils
Pillow>=9.0.0
tqdm
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.13.0
wandb
numpy>=1.24.0
datasets
einops
flash-attn ; platform_system == "Linux"
